Use in-tree NSS by reverting BeOS port removal and applying haikuports patchset

+++ security/nss/cmd/httpserv/httpserv.c
@@ -1321,7 +1321,7 @@ main(int argc, char **argv)
         FILE *tmpfile = fopen(pidFile, "w+");
 
         if (tmpfile) {
-            fprintf(tmpfile, "%d", getpid());
+            fprintf(tmpfile, "%d", (int)getpid());
             fclose(tmpfile);
         }
     }
